Combat Code
This is a script for a basic M1 and M2 combat where M2 is a heavy hit. This revolves around a system where fighting a shadow is considered "Training" which is the reason for a BP and LF value listed at the top. You are free to remove values and anything involved with them as that was just the direction I was going.